EnzoProlong updates
Control

  - bug fix: Moved deallocating fluxes from MethodFluxCorrect.cpp to
    control_compute.cpp to avoid memory leak if MethodFluxCorrect isn't
    called

Debug

  - Updating MethodDebug for added parameters (incomplete)
  - Added parameters to "debug" method: "print", "coarse", "ghost"
    (incomplete)
  - added debug code for copying fields in EnzoSolverBiCgStab

Parameters

  - Added method_refresh_prolong

Method

  - EnzoMethodGravity updates
    - Added index_prolong parameter to EnzoMethodGravity
    - Re-added accelerations to ir_post refresh in EnzoMethodGravity
    - fixed typo: naming ir_exit_ refresh

Solver

  - EnzoSolverBiCgStab can produce poor solutions at mesh level jumps
    if restart != 1: resetting to 1 with WARNING message to bypass
    until this is addressed

Cleaning

  - Removed obsolete EnzoMethodCheckGravity
  - Removed obsolete files in tools directory
  - Added tools/index.org file to document remaining available tools

+++ src/Cello/control_output.cpp
@@ -87,7 +87,7 @@ void Problem::new_output_next(Simulation * simulation) throw()
   // print blocks on this process to check block_vec_ array
 
   Hierarchy * hierarchy = simulation->hierarchy();
-  for (size_t i=0; i<hierarchy->num_blocks(); i++) {
+  for (int i=0; i<hierarchy->num_blocks(); i++) {
     CkPrintf ("%d Block %d = %s\n",CkMyPe(),i,
 	      hierarchy->block(i)->name().c_str());
   }
